requirements.txt
streamlit
google-generativeai
python-dotenv
base64